:sparkles: Add process payment use case to in ports.

+++ payment-gateway-adapter-acquiring-bank/src/main/java/paymentgateway/adapter/acquiringbank/AuthorizePaymentRequest.java
@@ -1,12 +1,11 @@
 package paymentgateway.adapter.acquiringbank;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import java.math.BigInteger;
 import lombok.Builder;
 
 @Builder
 record AuthorizePaymentRequest(@JsonProperty("card_number") String cardNumber,
                                @JsonProperty("expiry_date") String expiryDate, String currency,
-                               BigInteger amount, String cvv) {
+                               long amount, String cvv) {
 
 }
